Wrap new BuildGenerator in a lazy factory callable to prevent side-effects
Reported issue to: https://github.com/browscap/browscap/issues/2315

In the `__construct` methods of `WriterInterface` implementors, all
implementations I found produce side-effects, because:

```php
$ressource    = fopen($file, 'w');
```

This has the effect of resetting the file to a blank empty file as soon as the
implementor is instantiated:

```php
$iniWriter = new IniWriter($pathToFile, $logger);
```

Constructors should NOT have side-effects, and this is the source of a long-
standing bug I've had in `browscap/browscap-site` when checking to see if the
generated files are up to date (it would delete the files, but the metadata
would remain). For now, I've worked around it by creating the generator in a
lazy factory (just a `callable` that is invoked when the generator is actually
used to generate new files).